pyproject.toml
[tool.poetry]
name = "flame-api"
version = "0.1.0"
description = ""
authors = ["Denis <postnikovda@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
uvicorn = "^0.11.5"
fastapi = "^0.58.1"
pydantic = {extras = ["email"], version = "^1.5.1"}
motor = "^2.1.0"
firebase_admin = "^4.3.0"
databases = "^0.3.2"
pyjwt = "^1.7.1"
passlib = "^1.7.2"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"
fab-classic = "^1.18.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
